requirements.txt
Flask
Flask-SQLAlchemy
psycopg2-binary
python-dotenv
pdf2docx
pypdf
pikepdf
